Restrict Target JVM version detection code to plugins only
- Adds a second describer for non-plugin library requests.
- Adjusts testing and registration of describers in schemas as appropriate.
