+++ platforms/extensibility/plugin-use/src/main/java/org/gradle/plugin/use/internal/DefaultPluginArtifactRepositories.java
@@ -23,10 +23,8 @@ import org.gradle.api.internal.DocumentationRegistry;
 import org.gradle.api.internal.artifacts.DependencyResolutionServices;
 import org.gradle.api.internal.artifacts.JavaEcosystemSupport;
 import org.gradle.api.internal.artifacts.dsl.RepositoryHandlerInternal;
-import org.gradle.api.internal.artifacts.dsl.dependencies.TargetJVMVersionTooHighFailureDescriber;
 import org.gradle.api.internal.attributes.AttributesSchemaInternal;
 import org.gradle.internal.Factory;
-import org.gradle.internal.component.resolution.failure.type.IncompatibleGraphVariantFailure;
 import org.gradle.plugin.use.resolve.internal.ArtifactRepositoriesPluginResolver;
 import org.gradle.plugin.use.resolve.internal.PluginArtifactRepositories;
 import org.gradle.plugin.use.resolve.internal.PluginResolver;
@@ -44,7 +42,6 @@ class DefaultPluginArtifactRepositories implements PluginArtifactRepositories {
 
         AttributesSchemaInternal attributesSchema = (AttributesSchemaInternal) Objects.requireNonNull(dependencyResolutionServices).getAttributesSchema();
         JavaEcosystemSupport.configureSchema(attributesSchema, dependencyResolutionServices.getObjectFactory());
-        attributesSchema.addFailureDescriber(IncompatibleGraphVariantFailure.class, TargetJVMVersionTooHighFailureDescriber.class);
 
         RepositoryHandler repositoryHandler = dependencyResolutionServices.getResolveRepositoryHandler();
         for (ArtifactRepository repository : sharedRepositories) {
