+++ platforms/software/dependency-management/src/test/groovy/org/gradle/plugin/devel/variants/TargetJVMVersionOnLibraryTooNewFailureDescriberTest.groovy
@@ -1,0 +1,5 @@
+package org.gradle.plugin.devel.variants
+
+@interface TargetJVMVersionOnLibraryTooNewFailureDescriberTest {
+
+}
